+++ docs/AUTHORS.md
@@ -5,14 +5,15 @@ Development Lead
 ----------------
 
 -   Guillaume Viejo <gviejo@flatironinstitute.org>
+- 	Edoardo Balzani <ebalzani@flatironinstitute.org>
 
 
 Contributors
 ------------
 
-- 	Edoardo Balzani <ebalzani@flatironinstitute.org>
 -	Adrien Peyrache
 - 	Dan Levenstein
 -	Sofia Skromne Carrasco
 -	Davide Spalla
 - 	Luigi Petrucco
+  - ... [and many more!](https://github.com/pynapple-org/pynapple/graphs/contributors)
