+++ docs/examples/tutorial_phase_preferences.py
@@ -154,7 +154,7 @@ ax1.margins(0)
 # Filtering Theta
 # ---------------
 #
-# As expected, there is a strong 8Hz component during REM sleep. We can filter it using the function `nap.compute_bandpass_filter`.
+# As expected, there is a strong 8Hz component during REM sleep. We can filter it using the function `nap.apply_bandpass_filter`.
 
 theta_band = nap.apply_bandpass_filter(tsd_rem, cutoff=(6.0, 10.0), fs=FS)
