+++ pynapple/process/__init__.py
@@ -22,11 +22,9 @@ from .randomize import (
     shift_timestamps,
     shuffle_ts_intervals,
 )
-from .signal_processing import (
+from .spectrum import (
     compute_mean_power_spectral_density,
     compute_power_spectral_density,
-    compute_wavelet_transform,
-    generate_morlet_filterbank,
 )
 from .tuning_curves import (
     compute_1d_mutual_info,
@@ -37,3 +35,4 @@ from .tuning_curves import (
     compute_2d_tuning_curves_continuous,
     compute_discrete_tuning_curves,
 )
+from .wavelets import compute_wavelet_transform, generate_morlet_filterbank
